Use custom property for card background color
Since the button's text color must be the same as the card's background
color, it is convenient to create a custom property with the latter.

A generic button class will style a default button that could be used
even outside of the cards, and a card__button class will make the button
use the card's background color as its text color via the custom
property.